rename table_parquet_options.rs to file_format_options.rs
