+++ crates/core/src/operations/delete.rs
@@ -62,7 +62,7 @@ use crate::operations::CustomExecuteHandler;
 use crate::protocol::DeltaOperation;
 use crate::table::config::TablePropertiesExt as _;
 use crate::table::state::DeltaTableState;
-use crate::table::table_parquet_options::{
+use crate::table::file_format_options::{
     build_writer_properties_factory_ffo, build_writer_properties_factory_wp,
     state_with_file_format_options, to_table_parquet_options_from_ffo, FileFormatOptions,
     WriterPropertiesFactory,
