+++ crates/core/src/operations/drop_constraints.rs
@@ -10,7 +10,7 @@ use crate::kernel::{Action, MetadataExt};
 use crate::logstore::LogStoreRef;
 use crate::protocol::DeltaOperation;
 use crate::table::state::DeltaTableState;
-use crate::table::table_parquet_options::FileFormatOptions;
+use crate::table::file_format_options::FileFormatOptions;
 use crate::DeltaTable;
 use crate::{DeltaResult, DeltaTableError};
