+++ crates/core/src/operations/merge/mod.rs
@@ -93,7 +93,7 @@ use crate::operations::write::WriterStatsConfig;
 use crate::protocol::{DeltaOperation, MergePredicate};
 use crate::table::config::TablePropertiesExt as _;
 use crate::table::state::DeltaTableState;
-use crate::table::table_parquet_options::{
+use crate::table::file_format_options::{
     build_writer_properties_factory_ffo, build_writer_properties_factory_wp,
     state_with_parquet_options, to_table_parquet_options_from_ffo, FileFormatOptions,
     WriterPropertiesFactory,
