+++ crates/core/src/operations/optimize.rs
@@ -55,7 +55,7 @@ use crate::logstore::{LogStore, LogStoreRef, ObjectStoreRef};
 use crate::protocol::DeltaOperation;
 use crate::table::config::TablePropertiesExt as _;
 use crate::table::state::DeltaTableState;
-use crate::table::table_parquet_options::{
+use crate::table::file_format_options::{
     build_writer_properties_factory_ffo, build_writer_properties_factory_wp,
     to_table_parquet_options_from_ffo, FileFormatOptions, WriterPropertiesFactory,
 };
