+++ crates/core/src/operations/update.rs
@@ -59,7 +59,7 @@ use crate::logstore::LogStoreRef;
 use crate::operations::cdc::*;
 use crate::protocol::DeltaOperation;
 use crate::table::state::DeltaTableState;
-use crate::table::table_parquet_options::{
+use crate::table::file_format_options::{
     build_writer_properties_factory_ffo, build_writer_properties_factory_wp,
     state_with_parquet_options, to_table_parquet_options_from_ffo, FileFormatOptions,
     WriterPropertiesFactory,
