+++ crates/core/src/operations/write/mod.rs
@@ -76,7 +76,7 @@ use crate::kernel::{
 use crate::logstore::LogStoreRef;
 use crate::protocol::{DeltaOperation, SaveMode};
 use crate::table::state::DeltaTableState;
-use crate::table::table_parquet_options::{
+use crate::table::file_format_options::{
     build_writer_properties_factory_ffo, build_writer_properties_factory_wp, FileFormatOptions,
     WriterPropertiesFactory,
 };
