+++ crates/core/src/operations/write/writer.rs
@@ -22,7 +22,7 @@ use crate::crate_version;
 use crate::errors::{DeltaResult, DeltaTableError};
 use crate::kernel::{Add, PartitionsExt};
 use crate::logstore::ObjectStoreRef;
-use crate::table::table_parquet_options::{
+use crate::table::file_format_options::{
     build_writer_properties_factory_wp, WriterPropertiesFactory,
 };
 use crate::writer::record_batch::{divide_by_partition_values, PartitionResult};
