+++ crates/core/src/writer/json.rs
@@ -26,7 +26,7 @@ use crate::kernel::{scalars::ScalarExt, Add, PartitionsExt};
 use crate::logstore::ObjectStoreRetryExt;
 use crate::table::builder::DeltaTableBuilder;
 use crate::table::config::TablePropertiesExt as _;
-use crate::table::table_parquet_options::{
+use crate::table::file_format_options::{
     build_writer_properties_factory_or_default_ffo, WriterPropertiesFactory,
 };
 use crate::writer::utils::ShareableBuffer;
